fix baseTenant

+++ backend/src/main/java/com/chatbot/core/billing/account/service/BillingAccountService.java
@@ -190,7 +190,7 @@ public class BillingAccountService {
 
     private BillingAccountResponse mapToResponse(BillingAccount account) {
         return BillingAccountResponse.builder()
-                .id(account.getId())
+                .id((Long) account.getId())
                 .accountNumber(account.getAccountNumber())
                 .billingType(account.getBillingType())
                 .accountName(account.getAccountName())
